CXPackage.go: Collect the definition of CXPackage and all methods into one file.
